+++ CHANGELOG.rst
@@ -66,6 +66,15 @@ in development
 * Record failures to enforce rules due to missing actions or parameter validation errors. A
   RuleEnforcement object will be created for failed enforcements that do not lead to an
   ActionExecution creation. (improvement)
+* Add support for better serialization of the following parameter types for positional parameters
+  used in the local and remote script runner actions: ``integer``, ``float``, ``boolean``,
+  ``list``, ``object``. Previously those values were serialized as Python literals which made
+  parsing them in the shell scripts very cumbersome. Now they are serialized based on the simple
+  rules described in the documentation which makes it easy to use just by using simple shell
+  primitives such as if statements and ``IFS`` for lists. (improvement, new feature)
+* Fix ``linux.traceroute`` action. (bug fix)
+* Fix a bug with positional argument handling in the local script runner. Now the arguments with a
+  no value or value of ``None`` are correctly passed to the script. (bug fix)
 
 1.3.2 - February 12, 2016
 -------------------------